+++ src/LanguageSwitcher.jsx
@@ -1,13 +1,14 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 
-const LanguageSwitcher = () => {
+const LanguageSwitcher = ({ onCloseMenu }) => {
   const { t, i18n } = useTranslation();
 
   const changeLanguage = (lng) => {
     i18n.changeLanguage(lng);
+    onCloseMenu(); // Call the function passed from the Navbar to close the menu
   };
-  
+
   return (
     <div className="language-switcher gradient-bg">
         <h1>{t('LanguageSwitcher.info', 'Default text')}</h1>