+++ src/ScoreCounter.jsx
@@ -23,6 +23,9 @@ function App() {
   const [palletRate, setPalletRate] = useState(() => {
     return Number(localStorage.getItem("palletRate")) || 0;
   });
+  const [weightedRate, setWeightedRate] = useState(() => {
+    return Number(localStorage.getItem("weightedRate")) || 0;
+  })
   const [isAddingPause, setIsAddingPause] = useState(() => {
     return (localStorage.getItem("isAddingPause")) === "true";
   });
@@ -32,9 +35,10 @@ function App() {
   const [remaining, setRemaining] = useState(() => {
     return Number(localStorage.getItem('remaining')) || 15;
   });
-
-
-
+  const [isDebugging, setIsDebugging] = useState(false);
+  const [calculations, setCalculations] = useState([]);  // State to store the list of saved calculations
+  const [trucks, setTrucks] = useState([]);
+  const [activities, setActivities] = useState([]);  // State to store the list of saved activities
   const { t } = useTranslation();
 
   const startOver = () => {
@@ -43,7 +47,8 @@ function App() {
     setTotalPallets(0);
     setStartTime(null);
     setElapsedTime(0);
-    setPalletRate(0);
+    // setPalletRate(0);
+    setWeightedRate(0);
     setInputPallets('');
     setIsCounting(false);
     setStartTimeInput('');
@@ -55,7 +60,8 @@ function App() {
     // Clear localStorage
     localStorage.removeItem('totalPallets');
     localStorage.removeItem('elapsedTime');
-    localStorage.removeItem('palletRate');
+    // localStorage.removeItem('palletRate');
+    localStorage.removeItem('weightedRate');
     localStorage.removeItem('startTime');
     localStorage.removeItem('isCounting');
     localStorage.removeItem('isAddingPause');
@@ -63,40 +69,97 @@ function App() {
     localStorage.removeItem('pausedFor')
   };
 
-    // Persist states in localStorage
-    useEffect(() => {
-      localStorage.setItem("totalPallets", totalPallets);
-    }, [totalPallets]);
-  
-    useEffect(() => {
-      localStorage.setItem("elapsedTime", elapsedTime);
-    }, [elapsedTime]);
-  
-    useEffect(() => {
-      localStorage.setItem("palletRate", palletRate);
-    }, [palletRate]);
-  
-    useEffect(() => {
-      localStorage.setItem("isCounting", isCounting);
-    }, [isCounting]);
+  // Persist states in localStorage
+  useEffect(() => {
+    localStorage.setItem("totalPallets", totalPallets);
+  }, [totalPallets]);
 
-    useEffect(() => {
-      localStorage.setItem("isAddingPause", isAddingPause);
-    }, [isAddingPause]);
+  useEffect(() => {
+    localStorage.setItem("elapsedTime", elapsedTime);
+  }, [elapsedTime]);
+
+  useEffect(() => {
+    localStorage.setItem("palletRate", palletRate);
+  }, [palletRate]);
+
+  useEffect(() => {
+    localStorage.setItem("weightedRate", weightedRate);
+    }, [weightedRate]);
+
+  useEffect(() => {
+    localStorage.setItem("isCounting", isCounting);
+  }, [isCounting]);
+
+  useEffect(() => {
+    localStorage.setItem("isAddingPause", isAddingPause);
+  }, [isAddingPause]);
+
+  useEffect(() => {
+    if (startTime !== null) {
+      localStorage.setItem("startTime", startTime);
+    }
+  }, [startTime]);
+
+  useEffect(() => {
+    localStorage.setItem("pausedFor", pausedFor);
+  }, [pausedFor]);
+
+  useEffect(() => {
+    localStorage.setItem("remaining", remaining);
+  }, [remaining]);
+
+  //START OF TODAY's ACTIVITY FUNCTION
+  const saveActivity = () => {
+      // Save the current activity as an object
+      const newActivities = {
+        truckNumber: activities.length + 1,
+        inputPallets,
+        timestamp: new Date().toLocaleString(), // You can add a timestamp for reference
+      };
   
-    useEffect(() => {
-      if (startTime !== null) {
-        localStorage.setItem("startTime", startTime);
-      }
-    }, [startTime]);
+      // Add the new calculation to the list (use the previous state to ensure it's updated correctly)
+      setActivities(prevActivities => [newActivities, ...prevActivities]); // Add at the beginning
+  };
+  //END
 
-    useEffect(() => {
-      localStorage.setItem("pausedFor", pausedFor);
-    }, [pausedFor]);
+  //START OF CALCULATION SAVER FUNCTION
+  const saveCalculation = () => {
+    const isConfirmed = window.confirm("Are you sure you want to save and stop Your calculation?");
+    if(isConfirmed){
+      // Save the current calculation as an object
+      const newCalculation = {
+        palletRate: palletRate.toFixed(2),
+        elapsedTime,
+        totalPallets,
+        pausedFor,
+        timestamp: new Date().toLocaleString(), // You can add a timestamp for reference
+      };
+  
+      // Add the new calculation to the list (use the previous state to ensure it's updated correctly)
+      setCalculations(prevCalculations => [newCalculation, ...prevCalculations]); // Add at the beginning
+      startOver();
+    }
+  };
 
-    useEffect(() => {
-      localStorage.setItem("remaining", remaining);
-    }, [remaining]);
+  const clearCalculations = () => {
+    const isConfirmed = window.confirm("Are you sure you want to clear all calculations?");
+    if (isConfirmed) {
+      // Reset the calculations state to an empty array
+      setCalculations([]);
+      setActivities([]);
+    }
+  };
+
+  const formatElapsedTime = (elapsedTimeInSeconds) => {
+    const hours = Math.floor(elapsedTimeInSeconds / 3600);
+    const minutes = Math.floor((elapsedTimeInSeconds % 3600) / 60);
+    const seconds = Math.floor(elapsedTimeInSeconds % 60);
+
+    // Pad hours, minutes, and seconds to be always 2 digits
+    return `${hours.toString().padStart(2, '0')}:${minutes.toString().padStart(2, '0')}:${seconds.toString().padStart(2, '0')}`;
+  };
+  //END OF CALCULATION SAVER FUNCTION
+
   // Start counting function (sets the start time and initializes the pallets)
   const startCounting = () => {
 
@@ -117,43 +180,24 @@ function App() {
   // Function to add additional pallets
   const addPallets = () => {
     const newPallets = Number(inputPallets);
+    saveActivity();
     if (newPallets > 0) {
       setTotalPallets((prevTotal) => prevTotal + newPallets); // Use functional update to ensure latest state
       setInputPallets(''); // Clear the input field after submitting
     }
   };
 
-  // HERE SUBSTITUTE THIS FOR NEW PAUSE FUNCTION
-  // useEffect(() => {
+  // START OF DEBUGGING FUNCTIONS
+  const toggleDebugging = () => {
+    setIsDebugging(prevState => !prevState); // Toggle debugging state
+  };
 
-  //   if (paused) {
-  //     // Save the time when the pause starts
-  //     if (!pauseStartRef.current) {
-  //       pauseStartRef.current = Date.now();
-  //     }
+  const handlePauseReset = () => {
+    setRemaining(15);
+    setMessage("");
+    setPausedFor(0);
+  }
 
-  //     // Start the interval to update pauseTimer every second
-  //     interval = setInterval(() => {
-  //       const elapsed = Math.floor((Date.now() - pauseStartRef.current) / 1000); // Elapsed time in seconds
-  //       setPauseTimer((prevTimer) => prevTimer + 1); // Increment by 1 each second
-  //       pauseStartRef.current += 1000; // Update pauseStartRef to avoid cumulative additions
-  //     }, 1000);
-  //   } else {
-  //     // Add the final elapsed time when unpausing
-  //     if (pauseStartRef.current) {
-  //       const elapsed = Math.floor((Date.now() - pauseStartRef.current) / 1000); // Total elapsed time
-  //       setPauseTimer((prevTimer) => prevTimer + elapsed); // Add elapsed time once
-  //       pauseStartRef.current = null; // Reset pauseStartRef
-  //     }
-
-  //     // Clear the interval
-  //     clearInterval(interval);
-  //   }
-
-  // }, [paused]);
-
-
-  // Button for debugging purposes
   const handleDebugClick = () => {
     console.clear();
     console.log ('============================');
@@ -167,8 +211,15 @@ function App() {
     console.log ("Paused for:", pausedFor);
     console.log (Date.now() / 1000);
     console.log ("Remaining: ", remaining);
+    console.log ("IsDebugging?: ", isDebugging);
+    console.log ("Weighted score: ", weightedRate);
+    calculations.forEach((calc, index) => {
+      console.log(`  palletRate: ${calc.palletRate}`);
+      console.log(`  elapsedTime: ${calc.elapsedTime}`);
+    });
     console.log ('============================');
   }
+  // END OF DEBUGGING FUNCTIONS
 
   // START OF PAUSE FUNCTIONS
   const handleInputChange = (e) => {
@@ -183,7 +234,6 @@ function App() {
   const handlePauseClick = () => {
     setIsAddingPause(true); // Show the form
   }
-
 
   const handleSubmit = (e) => {
     e.preventDefault();
@@ -209,13 +259,38 @@ function App() {
 
         // Calculate pallets per hour
         const rate = (totalPallets / (currentElapsedTime - (pausedFor * 60))) * 3600;
-        setPalletRate(rate); // Update the rate of pallets per hour
+
+        if (rate > 0) {
+          // Check if there are previous calculations for weighted rate
+          if (calculations.length > 0){
+            const totalWeightedRate = calculations.reduce(
+              (acc, calc) => acc + calc.palletRate * calc.elapsedTime,
+              0
+            );
+            const totalElapsedTime = calculations.reduce(
+              (acc, calc) => acc + calc.elapsedTime,
+              0
+            );
+
+            // Calculate the weighted rate
+            const newWeightedRate =
+              (totalWeightedRate + rate * currentElapsedTime) /
+              (totalElapsedTime + currentElapsedTime);
+            console.log(totalWeightedRate, '+', rate, '*', currentElapsedTime, '/', totalElapsedTime, '+', currentElapsedTime);
+            // console.log(rate);
+            // console.log(currentElapsedTime);
+            // console.log(totalElapsedTime);
+            // console.log(currentElapsedTime);
+            setWeightedRate(newWeightedRate); // Update the weighted rate
+          } else {
+            setPalletRate(rate); // Update the rate of pallets per hour
+          }
+        }
       }, 100); // Update every 100ms
-  
       // Cleanup interval when pausing or stopping
       return () => clearInterval(intervalId);
     }
-  }, [isCounting, startTime, totalPallets, pausedFor]);
+  }, [isCounting, startTime, totalPallets, pausedFor, calculations]);
 
   return (
     <div className="score-counter">
@@ -224,18 +299,99 @@ function App() {
       {/* Display the current status only if counting has started*/}
       <div className="score-display">
         <hr className="ct"/>
-        <h2>{t('ScoreCounter.palam')}{totalPallets}</h2>
-        <h3>{t('ScoreCounter.palperh')}{palletRate.toFixed(2)}</h3>
+        <div className='statistics'>
+          <div className='stat-item'>
+            <span>{t('ScoreCounter.palperh')}</span>
+            <strong>
+              <span>{palletRate.toFixed(2)} pal/h</span>
+            </strong>
+          </div>
+          <div className='stat-item'>
+            <span>{t('ScoreCounter.palam')}</span>
+            <strong>
+              <span>{totalPallets} pal</span>
+            </strong>
+          </div>
+          <div className='stat-item'>
+            <span>{t('ScoreCounter.eltim')} </span>
+            <strong>
+              <span>
+                  {Math.floor(elapsedTime / 3600)}h {Math.floor((elapsedTime % 3600) / 60)}m {Math.floor(elapsedTime % 60)}s
+              </span>
+            </strong>
+          </div>
+          <div className='stat-item'>
+            <span>{t('ScoreCounter.elptim')}</span>
+            <strong>
+              <span>{Math.floor(pausedFor)} min</span>
+            </strong>
+          </div>
+        </div>
+        <hr className='ct'/>
+        <div className="statistics">
+          <div className="stat-item">
+            <span>Weighted score:</span>
+            <strong>
+              <span> {weightedRate.toFixed(2)} pal/h</span>
+            </strong>
+          </div>
+          <div className="stat-item">
+            <span>Total pallets loaded:</span>
+            <strong>
+              <span> (WIP) pal</span>
+            </strong>
+          </div>
+          <div className="stat-item">
+            <span>Current elapsed time:</span>
+            <strong>
+              <span> (WIP)</span>
+            </strong>
+          </div>
+        </div>
+        <hr className='ct'/>
+        <p>- Short statistics -</p>
         <p>
-          {t('ScoreCounter.eltim')} {Math.floor(elapsedTime / 3600)}h {Math.floor((elapsedTime % 3600) / 60)}m {Math.floor(elapsedTime % 60)}s
-        </p>
-        <p>
-          {t('ScoreCounter.elptim')}
-          {Math.floor(pausedFor)}min
+          <span>[Score: {palletRate.toFixed(2)}] </span>
+          <span>[Pal: {totalPallets}] </span>
+          <span>
+            [Time: {Math.floor(elapsedTime / 3600)}h {Math.floor((elapsedTime % 3600) / 60)}m {Math.floor(elapsedTime % 60)}s]
+          </span>
         </p>
         <hr className="ct"/>
-        </div>
-      
+      </div>
+
+      {/* Display the saved score after finishing calculations */}
+      <div>
+        <p>- Calculation Results -</p>
+        <ul>
+          {calculations.map((calc, index) => (
+            <li key={index}>
+              <strong>Calculation #{index + 1} ({calc.timestamp})</strong><br />
+              Rate: {calc.palletRate} pallets/hour<br />
+              Elapsed time: {formatElapsedTime(calc.elapsedTime)}<br />
+              Total pallets: {calc.totalPallets}<br />
+              Paused for: {calc.pausedFor} minutes
+            </li>
+          ))}
+        </ul>
+        <hr className='ct'/>
+      </div>
+
+      {/* Display the array of every truck loaded with amount of pallets and timestamp */}
+      <div>
+        <p>- Today's activity -</p>
+        <ul>
+          {activities.map((a, index) => (
+            <li key={index}>
+              <strong>Truck #{a.truckNumber} ({a.timestamp})</strong><br />
+              Pallets loaded: {a.inputPallets}<br />
+              Work efficiency: (WIP)
+              Loading time: (WIP)<br />
+            </li>
+          ))}
+        </ul>
+        <hr className='ct'/>
+      </div>
 
       {!isCounting ? (
         // Input for the start time
@@ -256,11 +412,19 @@ function App() {
           <button onClick={startCounting} disabled={startTimeInput === ''}>
             {t('ScoreCounter.buttons.startct')}
           </button>
-          <button onClick={startOver}>DEBUG-RESET</button>
-          {/* DEBUG BUTTON */}
-            <button onClick={handleDebugClick}>
-              Console.log DEBUG
-            </button>
+          <button className="debug" onClick={toggleDebugging}>
+            {isDebugging ? "Disable Debugging" : "Enable Debugging"}
+          </button>
+          {isDebugging && (
+            <div>
+              <button className="debug" onClick={startOver}>DEBUG-RESET</button>
+              <button className="debug" onClick={handleDebugClick}>
+                Console.log DEBUG
+              </button>
+              <button className="debug" onClick={clearCalculations}>Clear Calculations</button>
+              <button className="debug" onClick={() => setPalletRate(0)}>Clear Rate</button>
+            </div>
+          )}
         </div>
       ) : (
         // After starting, show the option to add additional pallets
@@ -280,54 +444,54 @@ function App() {
           <button onClick={addPallets}>
           {t('ScoreCounter.buttons.updtscr')}
           </button>
+          {/* ADD PAUSE BUTTON */}
+          {message ? (
+            <p>{message}</p>
+          ) : (
+            <>
+              {!isAddingPause && isCounting &&remaining > 0 && (
+                    <button onClick={handlePauseClick}>Add Pause</button>
+                  )}
+                  {isAddingPause && (
+                    <form onSubmit={handleSubmit}>
+                      <label>
+                        Choose a number (1-{remaining}):
+                        <input
+                          type="number"
+                          min="1"
+                          max={remaining}
+                          value={inputValue}
+                          onChange={handleInputChange}
+                          disabled={remaining === 0}
+                        />
+                      </label>
+                      <button type="submit" disabled={remaining === 0}>
+                        Submit
+                      </button>
+                    </form>
+                  )}
+            </>
+          )}
+          <button onClick={saveCalculation}>Save Calculation</button>
           <button onClick={startOver}>
           {t('ScoreCounter.buttons.startovr')}
           </button>
-          {/* Show the Pause/Resume button only if counting has started
-          {isCounting && (
-            <button 
-              onClick={handlePauseClick} // Toggle paused state
-              disabled={pausedFor === 15}
-            >
-              START THE PAUSE
-            </button>
-          )} */}
           {/* DEBUG BUTTON */}
-          {isCounting && (
-            <button onClick={handleDebugClick}>
-              Console.log DEBUG
-            </button>
+          <button className="debug" onClick={toggleDebugging}>
+            {isDebugging ? "Disable Debugging" : "Enable Debugging"}
+          </button>
+          {isDebugging && (
+            <div>
+              <button className="debug" onClick={handleDebugClick}>
+                Console.log DEBUG
+              </button>
+              <button className="debug" onClick={handlePauseReset}>
+                Reset Pause
+              </button>
+              <button className="debug" onClick={clearCalculations}>Clear Calculations</button>
+            </div>
           )}
         </div>
-      )}
-
-      {/* ADD PAUSE BUTTON */}
-      {message ? (
-        <p>{message}</p>
-      ) : (
-        <>
-          {!isAddingPause && remaining > 0 && (
-                <button onClick={handlePauseClick}>Add Pause</button>
-              )}
-              {isAddingPause && (
-                <form onSubmit={handleSubmit}>
-                  <label>
-                    Choose a number (1-{remaining}):
-                    <input
-                      type="number"
-                      min="1"
-                      max={remaining}
-                      value={inputValue}
-                      onChange={handleInputChange}
-                      disabled={remaining === 0}
-                    />
-                  </label>
-                  <button type="submit" disabled={remaining === 0}>
-                    Submit
-                  </button>
-                </form>
-              )}
-        </>
       )}
 
         {/*HOW-TO SECTION*/}